test(citizen): keep source guards Node-only so Cypress CT can run
The citizen profile page metadata guards used fs.readFileSync at describe
load time. That works under Mocha in Node, but Cypress component tests run
in the browser where fs is stubbed without readFileSync — which is what
failed component-tests (4) on CI.

Move those guards into a Node-only mocha script and leave the pure unit
assertions in the Cypress/Mocha shared spec.
