+++ ui/cypress/integration/unit/new-citizen-discord-notification.cy.ts
@@ -1,6 +1,4 @@
 import { IPFS_GATEWAY } from 'const/config'
-import fs from 'fs'
-import path from 'path'
 import {
   buildNewCitizenBody,
   buildNewCitizenContent,
@@ -37,37 +35,6 @@ describe('citizen profile og:image', () => {
 
   it('leaves non-IPFS absolute URLs alone', () => {
     expect(normalizeOgImageUrl('https://example.com/a.png')).to.equal('https://example.com/a.png')
-  })
-})
-
-describe('citizen profile page metadata', () => {
-  const source = fs.readFileSync(
-    path.join(__dirname, '../../../pages/citizen/[tokenIdOrName].tsx'),
-    'utf8'
-  )
-
-  // This is the defect that kept coming back: normalizeOgImageUrl already routed
-  // ipfs:// to the working gateway, but the page pre-built an ipfs.io URL, which
-  // reached the helper looking like an already-resolved https URL and was passed
-  // straight through to Discord.
-  it('does not hardcode a public IPFS gateway for the portrait', () => {
-    for (const gateway of [
-      'https://ipfs.io/ipfs/',
-      'https://dweb.link/ipfs/',
-      'https://cloudflare-ipfs.com/ipfs/',
-      'https://gateway.pinata.cloud/ipfs/',
-    ]) {
-      expect(source).to.not.include(gateway)
-    }
-  })
-
-  it('hands the raw metadata image to Head so it gets normalized', () => {
-    expect(source).to.include('image={nft?.metadata?.image}')
-  })
-
-  // A citizen with no portrait used to yield `https://ipfs.io/ipfs/undefined`.
-  it('does not split the portrait URI into a gateway path', () => {
-    expect(source).to.not.include("image.split('ipfs://')")
   })
 })
